+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 
+
+### Fixed
+- Do not close `ToggleButton` if `target`'s child is clicked during click outside event
+
+### Changed
+- Always close a toggle button's target after navigation
+
 ## [0.4.0] - 2022-10-13
 
 ### Changed